feat: Implement ragged (CSR) format for dataset caching and loading
- Introduced a new dataset cache format (v2.0) that supports ragged arrays, eliminating global padding.
- Updated `StageViewDataset` to handle both ragged and padded modes, maintaining backward compatibility with v1.x.
- Added `build_ragged_arrays` function to construct CSR arrays from per-walk numpy arrays.
- Modified data preparation and loading functions to accommodate the new format.
- Enhanced data loading with `BucketBatchSampler` for efficient batching of variable-length sequences.
- Updated tests to validate the new dataset format and ensure compatibility with existing functionality.

+++ src/model/lit_model.py
@@ -145,7 +145,7 @@ class LitEdgeClassifier(pl.LightningModule):
         if pool_edge_ids.numel() == 0:
             raise RuntimeError("dynamic_train_masking pool is empty")
 
-        pool_edge_classes = id2class[input_ids[pool_mask]].long().cpu()
+        pool_edge_classes = id2class[input_ids[pool_mask].long()].long().cpu()
         valid = pool_edge_classes != self.ignore_index
         if not valid.any():
             raise RuntimeError("dynamic_train_masking pool has no valid classes")
